combine transferui and link token create calls
