+++ server/routes/transfers.js
@@ -22,58 +22,6 @@ const router = express.Router();
 const { PLAID_CLIENT_ID, PLAID_SECRET_SANDBOX } = process.env;
 
 /**
- * creates a transfer intent for Transfer UI and retrieves a transfer_intent_id
- *
- * @param {string} subscriptionAmount the amount of the transfer.
- * @returns {Object}  transfer intent response
- */
-
-router.post(
-  '/transfer_ui',
-  asyncWrapper(async (req, res) => {
-    try {
-      const { userId, subscriptionAmount } = req.body;
-      const { username: username } = await retrieveUserById(userId);
-      const transIntentCreateRequest = {
-        client_id: PLAID_CLIENT_ID,
-        secret: PLAID_SECRET_SANDBOX,
-        mode: 'PAYMENT',
-        amount: subscriptionAmount.toFixed(2),
-        ach_class: 'ppd',
-        description: 'payment', // cannot be longer than 8 characters
-        user: {
-          legal_name: username,
-        },
-      };
-      let transferIntentId;
-
-      const transferIntentCreateResponse = await plaid.transferIntentCreate(
-        transIntentCreateRequest
-      );
-      transferIntentId = transferIntentCreateResponse.data.transfer_intent.id;
-      // create new Transfer now so that you can reference its transferIntentId upon link success
-      // because the link success metadata does not pass back any data about the transfer except for transfer_status
-      const newTransfer = await createTransfer(
-        null, // item_id
-        userId,
-        null, // plaid_account_id
-        transferIntentId,
-        null, // authorization_id - for TransferUI transfers
-        null, // transfer_id
-        subscriptionAmount.toFixed(2),
-        null, // status
-        'debit',
-        null // sweep_status
-      );
-      res.json(transferIntentCreateResponse.data);
-    } catch (err) {
-      console.log('error while creating transfer intent id', err.response.data);
-      return res.json(err.response.data);
-    }
-  })
-);
-
-/**
  * creates a transfer authorization for Transfer, retrieves an authorization_id to use to
  * create a transfer.
  *
